+++ New Thingy.csproj
@@ -1,0 +1,7 @@
+<Project Sdk="Godot.NET.Sdk/4.0.3">
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <EnableDynamicLoading>true</EnableDynamicLoading>
+    <RootNamespace>NewThingy</RootNamespace>
+  </PropertyGroup>
+</Project>
