+++ New Thingy.sln
@@ -1,0 +1,19 @@
+﻿Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2012
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "New Thingy", "New Thingy.csproj", "{A13A0A22-F4A3-43E4-90C3-9A87D47658CD}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+	Debug|Any CPU = Debug|Any CPU
+	ExportDebug|Any CPU = ExportDebug|Any CPU
+	ExportRelease|Any CPU = ExportRelease|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A13A0A22-F4A3-43E4-90C3-9A87D47658CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A13A0A22-F4A3-43E4-90C3-9A87D47658CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A13A0A22-F4A3-43E4-90C3-9A87D47658CD}.ExportDebug|Any CPU.ActiveCfg = ExportDebug|Any CPU
+		{A13A0A22-F4A3-43E4-90C3-9A87D47658CD}.ExportDebug|Any CPU.Build.0 = ExportDebug|Any CPU
+		{A13A0A22-F4A3-43E4-90C3-9A87D47658CD}.ExportRelease|Any CPU.ActiveCfg = ExportRelease|Any CPU
+		{A13A0A22-F4A3-43E4-90C3-9A87D47658CD}.ExportRelease|Any CPU.Build.0 = ExportRelease|Any CPU
+	EndGlobalSection
+EndGlobal
