--- game_level.tscn
@@ -1,8 +1,0 @@
-[gd_scene load_steps=2 format=3 uid="uid://c01c6viuyyu5j"]
-
-[ext_resource type="PackedScene" uid="uid://dr6fo4b2gdr3w" path="res://Characters/player_char.tscn" id="1_dcw2c"]
-
-[node name="game_level" type="Node"]
-
-[node name="PlayerChar" parent="." instance=ExtResource("1_dcw2c")]
-position = Vector2(266, 220)
