+++ levels/game_level.tscn
@@ -1,0 +1,79 @@
+[gd_scene load_steps=5 format=3 uid="uid://c01c6viuyyu5j"]
+
+[ext_resource type="Texture2D" uid="uid://dsgci0mkhbcbi" path="res://Assets/PlatformTiles.png" id="1_m0ppp"]
+[ext_resource type="PackedScene" uid="uid://dr6fo4b2gdr3w" path="res://Characters/player_char.tscn" id="2_1g183"]
+
+[sub_resource type="TileSetAtlasSource" id="TileSetAtlasSource_02jod"]
+texture = ExtResource("1_m0ppp")
+texture_region_size = Vector2i(10, 10)
+0:0/0 = 0
+0:0/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+0:0/0/physics_layer_0/angular_velocity = 0.0
+0:0/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+1:0/0 = 0
+1:0/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+1:0/0/physics_layer_0/angular_velocity = 0.0
+1:0/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+2:0/0 = 0
+2:0/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+2:0/0/physics_layer_0/angular_velocity = 0.0
+2:0/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+3:0/0 = 0
+3:0/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+3:0/0/physics_layer_0/angular_velocity = 0.0
+3:0/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+0:1/0 = 0
+0:1/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+0:1/0/physics_layer_0/angular_velocity = 0.0
+0:1/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+1:1/0 = 0
+1:1/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+1:1/0/physics_layer_0/angular_velocity = 0.0
+1:1/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+2:1/0 = 0
+2:1/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+2:1/0/physics_layer_0/angular_velocity = 0.0
+2:1/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+2:2/0 = 0
+2:2/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+2:2/0/physics_layer_0/angular_velocity = 0.0
+2:2/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+1:2/0 = 0
+1:2/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+1:2/0/physics_layer_0/angular_velocity = 0.0
+1:2/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+0:2/0 = 0
+0:2/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+0:2/0/physics_layer_0/angular_velocity = 0.0
+0:2/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+0:3/0 = 0
+0:3/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+0:3/0/physics_layer_0/angular_velocity = 0.0
+0:3/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+1:3/0 = 0
+1:3/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+1:3/0/physics_layer_0/angular_velocity = 0.0
+1:3/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+2:3/0 = 0
+2:3/0/physics_layer_0/linear_velocity = Vector2(0, 0)
+2:3/0/physics_layer_0/angular_velocity = 0.0
+2:3/0/physics_layer_0/polygon_0/points = PackedVector2Array(-5, -5, 5, -5, 5, 5, -5, 5)
+
+[sub_resource type="TileSet" id="TileSet_rexpx"]
+tile_size = Vector2i(10, 10)
+physics_layer_0/collision_layer = 1
+sources/6 = SubResource("TileSetAtlasSource_02jod")
+
+[node name="game_level" type="Node"]
+
+[node name="TileMap" type="TileMap" parent="."]
+tile_set = SubResource("TileSet_rexpx")
+format = 2
+layer_0/tile_data = PackedInt32Array(-1, 65542, 0, -65536, 65542, 0, -2, 6, 0, -65535, 131078, 0, 65534, 131078, 2, 65535, 65542, 3, 0, 65542, 3, 1, 131078, 3, 196601, 6, 0, 196602, 65542, 0, 196603, 65542, 0, 196604, 65542, 0, 196605, 65542, 0, 196606, 65542, 0, 196607, 65542, 0, 131072, 65542, 0, 131073, 65542, 0, 131074, 65542, 0, 131075, 65542, 0, 131076, 65542, 0, 131077, 65542, 0, 131078, 131078, 0, 196614, 65542, 2, 262150, 131078, 3, 262149, 65542, 3, 262148, 65542, 3, 262147, 65542, 3, 262146, 65542, 3, 262145, 65542, 3, 262144, 65542, 3, 327679, 65542, 3, 327678, 65542, 3, 327677, 65542, 3, 327676, 65542, 3, 327675, 65542, 3, 327674, 65542, 3, 327673, 131078, 2, 262137, 65542, 1, 262138, 196614, 0, 262139, 196614, 0, 262140, 196614, 0, 262141, 196614, 0, 262142, 196614, 0, 262143, 196614, 0, 196608, 196614, 0, 196609, 196614, 0, 196610, 196614, 0, 196611, 196614, 0, 196612, 196614, 0, 196613, 196614, 0)
+
+[node name="PlayerChar" parent="." instance=ExtResource("2_1g183")]
+position = Vector2(0, -31)
+
+[node name="Camera2D" type="Camera2D" parent="PlayerChar"]
+position = Vector2(0, 31)
+zoom = Vector2(5, 5)
